render_mode finished

+++ src/engine/render/render_ray_init.c
@@ -6,7 +6,7 @@
 /*   By: kearmand <kearmand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/10 13:41:08 by kearmand          #+#    #+#             */
-/*   Updated: 2025/12/10 13:41:25 by kearmand         ###   ########.fr       */
+/*   Updated: 2025/12/11 11:20:06 by kearmand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,15 +15,15 @@
 #include "engine.h"
 
 void	build_ray_for_pixel(const t_render_view *view,
-			int img_x, int img_y, t_ray *ray)
+			float img_x, float img_y, t_ray *ray)
 {
 	t_vec3f	offset_x;
 	t_vec3f	offset_y;
 	t_vec3f	pixel_pos;
 	t_vec3f	dir;
 
-	offset_x = vec3f_scale(view->dx, (float)img_x);
-	offset_y = vec3f_scale(view->dy, (float)img_y);
+	offset_x = vec3f_scale(view->dx, img_x);
+	offset_y = vec3f_scale(view->dy, img_y);
 	pixel_pos = vec3f_add(view->p0, offset_x);
 	pixel_pos = vec3f_add(pixel_pos, offset_y);
 	dir = vec3f_sub(pixel_pos, view->origin);
